web-app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
asyncssh==2.20.0
atpublic==5.1
attrs==25.3.0
beautifulsoup4==4.13.3
billiard==4.2.1
blinker==1.9.0
celery==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.9
cryptography==44.0.2
cssselect==1.3.0
datasets==3.5.0
dictdiffer==0.9.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dotenv==0.9.9
dpath==2.2.0
dulwich==0.22.8
dvc==3.59.1
dvc-data==3.16.10
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
dvclive==3.48.2
entrypoints==0.4

fastapi==0.115.12
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
filetype==1.2.0
Flask==3.1.0
flatten-dict==0.4.2
flufl.lock==8.1.0
frozenlist==1.6.0
fsspec==2024.12.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.44
grandalf==0.8
gto==1.7.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
hydra-core==1.3.2
idna==3.10

iterative-telemetry==0.0.10
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.4.2
Js2Py_3.13==0.74.1

kombu==5.5.3
lxml==5.3.2
lxml_html_clean==0.4.2
m3u8==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
propcache==0.3.1
psycopg2-binary==2.9.10

pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydot==3.0.4
pygit2==1.18.0
pygtrie==2.5.0
pyjsparser==2.7.1
pynvml==12.0.0
pyotp==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==14.0.0
ruamel.yaml==0.18.10
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
scmrepo==3.3.10
semver==3.0.4
sentence-transformers==4.0.2
setuptools==75.8.0
sgmllib3k==1.0.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
sqltrie==0.11.2
starlette==0.46.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.2.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.6.0
tqdm==4.67.1
transformers==4.51.1
triton==3.2.0
twikit==2.3.3
typer==0.15.3
typing-inspection==0.4.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.2
vine==5.1.0
voluptuous==0.15.2
webvtt-py==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.0
zc.lockfile==3.0.post1
